app/serenity_project/requirements.txt
asgiref==3.2.10
Django==3.1.1
djangorestframework==3.14.0
gunicorn==20.1.0
pytz==2022.2.1
sqlparse==0.4.3
whitenoise==6.2.0
